move work type logic to detective, and prioritize pubmed information
