+++ detective/__init__.py
@@ -1,0 +1,1 @@
+from .work_type_detective import WorkTypeDetective
